requirements.txt
python-docx>=1.1.2
openpyxl>=3.1.2
PyPDF2>=3.0.0
